simple parsing
I made simple parsing.
In the console, you need to enter the following string:
-t yes/no (working with the console or with a file) -i <input_file> -o <out_file>

My program checks the following cases:

If the user runs it from CodeBlocks, everything is fine.
If the user runs it without additional flags in the console, it works with the console.
If the user enters 'yes' and then enters anything else, the input is incorrect.
If the user makes a mistake in the order or names of the flags, the input is incorrect.
If the user enters files without extensions, the input is incorrect.
And most importantly:
If the user enters files that are not in .txt format, the program indicates that .txt is required.